requirements.txt
streamlit>=1.28.0
scikit-learn>=1.2.0
nltk>=3.8
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=1.0.0